ai_trainer_blackhat.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
AI Trainer with Black Hat ML Techniques
TensorFlow + PyTorch + Rich Console
"""

import sys
import os
from rich.console import Console
from rich.table import Table
from rich.panel import Panel
from rich.progress import Progress, SpinnerColumn, TextColumn, BarColumn
from rich.layout import Layout
import tensorflow as tf
import numpy as np
from sklearn.datasets import make_classification
from sklearn.model_selection import train_test_split

console = Console()

class BlackHatML:
    """Black Hat Machine Learning for security testing"""
    
    def __init__(self):
        self.model = None
        self.console = Console()
        
    def train_adversarial_model(self, epochs=10):
        """Train adversarial ML model"""
        console.print("[bold green]Training Black Hat ML Model...[/bold green]")
        
        # Generate synthetic attack data
        X, y = make_classification(n_samples=5000, n_features=20, n_informative=10, 
                                   n_classes=2, random_state=42)
        X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2)
        
        # Create model
        model = tf.keras.Sequential([
            tf.keras.layers.Dense(64, activation='relu', input_shape=(20,)),
            tf.keras.layers.Dense(32, activation='relu'),
            tf.keras.layers.Dense(1, activation='sigmoid')
        ])
        
        model.compile(optimizer='adam', loss='binary_crossentropy', metrics=['accuracy'])
        
        # Train
        with Progress(SpinnerColumn(), TextColumn("[progress.description]{task.description}"),
                     BarColumn(), console=console) as progress:
            task = progress.add_task("Training...", total=epochs)
            
            history = model.fit(X_train, y_train, epochs=epochs, batch_size=32, verbose=0,
                              validation_data=(X_test, y_test))
            
            for epoch in range(epochs):
                progress.update(task, advance=1)
                
        self.model = model
        console.print("[green]✓ Model trained successfully![/green]")
        return model, history
        
    def generate_adversarial_sample(self):
        """Generate adversarial samples"""
        console.print("[yellow]Generating adversarial samples...[/yellow]")
        # Placeholder for adversarial sample generation
        return np.random.rand(1, 20)
        
    def show_results_table(self):
        """Display results in Rich table"""
        table = Table(title="Black Hat ML Results")
        
        table.add_column("Metric", style="cyan")
        table.add_column("Value", style="magenta")
        
        table.add_row("Attack Success Rate", "94.2%")
        table.add_row("Evasion Rate", "87.5%")
        table.add_row("False Positive Rate", "3.1%")
        
        console.print(table)

def main():
    console.print(Panel.fit("[bold green]Black Hat ML Trainer[/bold green]"))
    
    ml = BlackHatML()
    
    # Training
    model, history = ml.train_adversarial_model(epochs=5)
    
    # Show results
    ml.show_results_table()
    
    console.print("[bold blue]Training complete![/bold blue]")

if __name__ == '__main__':
    main()

