requirements.txt
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
Flask==3.0.0
requests==2.31.0
scapy==2.5.0
python-whois==0.8.0
dnspython==2.4.2
rich==13.7.0
tensorflow==2.15.0
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
scikit-learn==1.3.2
pynput==1.7.6
pyaudio==0.2.14
requests==2.31.0
beautifulsoup4==4.12.2
qrcode==7.4.2
scapy==2.5.0
cryptography==41.0.7
Pillow==10.1.0
pandas==2.1.4

